+++ src/pharmpy/modeling/common.py
@@ -27,14 +27,50 @@ def read_model(path):
 
     Example
     -------
-    >>> read_model("/home/run1.mod")    # doctest: +SKIP
+    >>> from pharmpy.modeling import read_model
+    >>> model = read_model("/home/run1.mod")    # doctest: +SKIP
 
     See also
     --------
+    read_model_from_database : Read model from database
     read_model_from_string : Read model from string
 
     """
     model = Model(path)
+    return model
+
+
+def read_model_from_database(name, database=None):
+    """Read model from model database
+
+    Parameters
+    ----------
+    name : str
+        Name of model to use as lookup
+    database : Database
+        Database to use. Will use default database if not specified.
+
+    Returns
+    -------
+    Model
+        Read model object
+
+    Examples
+    --------
+    >>> from pharmpy.modeling import read_model_from_database
+    >>> model = read_model_from_database("run1")    # doctest: +SKIP
+
+    See also
+    --------
+    read_model : Read model from file
+    read_model_from_string : Read model from string
+
+    """
+    if database is None:
+        import pharmpy.workflows
+
+        database = pharmpy.workflows.default_model_database
+    model = database.get_model(name)
     return model
 
 
@@ -69,6 +105,7 @@ def read_model_from_string(code):
     See also
     --------
     read_model : Read model from file
+    read_model_from_database : Read model from database
 
     """
     model = Model(StringIO(code))